feat: learning outcomes
Learning outcomes can now be specified on a `section` using the `learningOutcomes` tag in the `yaml` header.

`learningOutcomes` should be a list of strings.

These are then rendered in an alert component below the title and before the content. Each learning outcome is prefixed with a :trophy: icon.

The list can be minimized by clicking the heading.