feat: Add docker-compose file
Adds a docker compose file which can be used for easily getting up and
running with a dev environment. Adds material pulling into the
Dockerfile with controllable behaviour via the MATERIAL_METHOD build-arg,
either 'copy' in a local folder or pull the repo described in a local
config yaml file. Also moves database migration to the entrypoint script
so that a working database is not required at build-time.