feat: Adds cache-busting for git pull
Adds support for cache-busting the git-pull of the material repo,
meaning that the pull section of the dockerfile will not be run
from the cached layer if the CACHE_BUST environment variable is
changed between calls to `docker build` or `docker-compose build`.

So far there seems no way to dynamically set the environment
variable from within the compose file, so a script is included to
do this manually via export, equally a prepended env variable could
be used when calling `docker-compose build` or `... up --build`.