Add a Docker compose file (#215)
* feat: Add docker-compose file

Adds a docker compose file which can be used for easily getting up and
running with a dev environment. Adds material pulling into the
Dockerfile with controllable behaviour via the MATERIAL_METHOD build-arg,
either 'copy' in a local folder or pull the repo described in a local
config yaml file. Also moves database migration to the entrypoint script
so that a working database is not required at build-time.

* fix: Remove wip comments from compose file

* fix: Refer to correct .material file with volume mount config

* fix: Ran prettier on README

* feat: Adds cache-busting for git pull

Adds support for cache-busting the git-pull of the material repo,
meaning that the pull section of the dockerfile will not be run
from the cached layer if the CACHE_BUST environment variable is
changed between calls to `docker build` or `docker-compose build`.

So far there seems no way to dynamically set the environment
variable from within the compose file, so a script is included to
do this manually via export, equally a prepended env variable could
be used when calling `docker-compose build` or `... up --build`.

* feat: Add nginx and qdrant to docker-compose

* move docker compose instructions to docs

* clean up nginx conf

* added workflow to test docker compose build

* sigh prettier

* removed mount material to allow pulled material in container

fixed some issues with the shell scripts

---------

Co-authored-by: Alasdair Wilson <alasdair.wlsn@googlemail.com>