fix: symlink libssl on latest Alpine (#318)
symlink `libssl` in the runner image on Alpine 3.21, as a workaround for older Prisma clients looking for it in `/lib`.

This fixes an issue for me where the Prisma client errors after building and running the latest Docker image. The app runs, but won't connect to the database.

```bash
gutenberg-1  | prisma:warn Prisma failed to detect the libssl/openssl version to use, and may not work as expected. Defaulting to "openssl-1.1.x".
gutenberg-1  | Please manually install OpenSSL and try installing Prisma again.
```