fix(a11y): event timeline buttons (#362)
* fix(a11y): event timeline buttons

Fix an accessibility bug where the Duplicate/Delete Event buttons are clickable SVG graphics, so can't be operated from the keyboard or with a screen reader. That's a WCAG level A failure, as those buttons couldn't be operated at all except by a sighted mouse/pointer user.

I've also refactored the event time to add a machine-readable `datetime` attribute. That's a minor change. I don't know if anyone supports that attribute, outside of maybe search engines like Google. I moved the machine-readable time inside the event link, but the flexbox layout should still be preserved for the time and admin buttons.

- add ARIA button roles and states to clickable event icons.

* Use tooltips as unique button labels